finana - growth valuation model for growth companies

+++ src/main/java/com/testehan/finana/model/valuation/growth/GrowthOutput.java
@@ -2,10 +2,8 @@ package com.testehan.finana.model.valuation.growth;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @Data
 public class GrowthOutput {
-    private BigDecimal intrinsicValuePerShare;
+    private Double intrinsicValuePerShare;
     private String verdict;
 }
